rename plugin to pass-thru
