+++ plugins/builtin/strategy/pass-thru/plugin/plugin_test.go
